#85 Decreased amount of sql-queries to DB per one http-query
