+++ Cooper.Repositories/UserRepository/IUserRepository.cs
@@ -1,9 +1,10 @@
 ﻿using Cooper.Models;
-
+using System.Collections.Generic;
 namespace Cooper.Repositories
 {
     public interface IUserRepository
     {
         User GetByJWToken(string token);
+        IList<User> GetUsersById(IList<long> UsersId);
     }
 }
